Split display form tests into input and actions file
