+++ src/components/__tests__/DisplayFormTests/DisplayActionButtons.test.jsx
@@ -1,0 +1,120 @@
+import { MemoryRouter } from 'react-router-dom';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import DisplayForm from '../../DisplayForm';
+import {
+  FIELD_TEXT,
+  SINGLE_PAGE_FORM,
+  VALIDATE_REQUIRED,
+} from '../../../constants/AppConstants';
+import { YOUR_VOYAGES_URL } from '../../../constants/AppUrlConstants';
+
+/*
+ * These tests check that we can pass a variety of
+ * form field objects, and form action objects
+ * to the display component
+ * and it will return HTML that will display a full form
+ * including labels, hints, inputs, buttons based on what we
+ * pass in
+ * It does NOT test a specific form structure or wording
+ * (that is done on the page that hold the specific form)
+ */
+
+const mockedUseNavigate = jest.fn();
+
+jest.mock('react-router-dom', () => ({
+  ...jest.requireActual('react-router-dom'),
+  useNavigate: () => mockedUseNavigate,
+}));
+
+describe('Display Form action buttons', () => {
+  const handleSubmit = jest.fn();
+  const scrollIntoViewMock = jest.fn();
+  window.HTMLElement.prototype.scrollIntoView = scrollIntoViewMock;
+  const formActions = {
+    submit: {
+      label: 'Submit test button',
+    },
+    cancel: {
+      label: 'Cancel test button',
+      redirectURL: YOUR_VOYAGES_URL,
+    },
+  };
+  const formActionsSubmitOnly = {
+    submit: {
+      label: 'Submit test button',
+    },
+  };
+  const formRequiredTextInput = [
+    {
+      type: FIELD_TEXT,
+      label: 'Text input',
+      hint: 'This is a hint for a text input',
+      fieldName: 'testField',
+      validation: [
+        {
+          type: VALIDATE_REQUIRED,
+          message: 'Enter your text input value',
+        },
+      ],
+    },
+  ];
+
+  beforeEach(() => {
+    window.sessionStorage.clear();
+  });
+
+  // ACTION BUTTONS
+  it('should render a submit and cancel button if both exist', () => {
+    render(
+      <MemoryRouter>
+        <DisplayForm
+          formId="testForm"
+          fields={formRequiredTextInput}
+          formActions={formActions}
+          formType={SINGLE_PAGE_FORM}
+          handleSubmit={handleSubmit}
+        />
+      </MemoryRouter>,
+    );
+    expect((screen.getByTestId('submit-button')).outerHTML).toEqual('<button type="button" class="govuk-button" data-module="govuk-button" data-testid="submit-button">Submit test button</button>');
+    expect((screen.getByTestId('cancel-button')).outerHTML).toEqual('<button type="button" class="govuk-button govuk-button--secondary" data-module="govuk-button" data-testid="cancel-button">Cancel test button</button>');
+  });
+
+  it('should render only a submit button if there is no cancel button', () => {
+    render(
+      <MemoryRouter>
+        <DisplayForm
+          formId="testForm"
+          fields={formRequiredTextInput}
+          formActions={formActionsSubmitOnly}
+          formType={SINGLE_PAGE_FORM}
+          handleSubmit={handleSubmit}
+        />
+      </MemoryRouter>,
+    );
+    expect(screen.getByTestId('submit-button').outerHTML).toEqual('<button type="button" class="govuk-button" data-module="govuk-button" data-testid="submit-button">Submit test button</button>');
+    expect(screen.getAllByRole('button')).toHaveLength(1);
+  });
+
+  it('should call handleSubmit function if submit button is clicked and there are no errors', async () => {
+    const user = userEvent.setup();
+    render(
+      <MemoryRouter>
+        <DisplayForm
+          formId="testForm"
+          fields={formRequiredTextInput}
+          formActions={formActionsSubmitOnly}
+          formType={SINGLE_PAGE_FORM}
+          handleSubmit={handleSubmit}
+        />
+      </MemoryRouter>,
+    );
+
+    await user.type(screen.getByLabelText('Text input'), 'Hello');
+    expect(screen.getByTestId('submit-button').outerHTML).toEqual('<button type="button" class="govuk-button" data-module="govuk-button" data-testid="submit-button">Submit test button</button>');
+    expect(screen.getAllByRole('button')).toHaveLength(1);
+    await user.click(screen.getByRole('button', { name: 'Submit test button' }));
+    expect(handleSubmit).toHaveBeenCalled();
+  });
+});
